+++ src/main/java/dev/orne/http/client/AbstractStatusIndependentUrlEncodedPutOperation.java
@@ -39,12 +39,7 @@ extends AbstractStatusIndependentPutOperation<P, E, R> {
     }
 
     /**
-     * Creates the HTTP request entity.
-     * 
-     * @param params The operation execution parameters
-     * @return The generated HTTP entity
-     * @throws HttpClientException If an exception occurs generating the
-     * entity
+     * {@inheritDoc}
      */
     @Override
     @Nullable